+++ app/src/main/java/dev/supergooey/pical/features/edit/EditLogFeature.kt
@@ -1,10 +1,10 @@
-package dev.supergooey.caloriesnap.features.edit
+package dev.supergooey.pical.features.edit
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import dev.supergooey.caloriesnap.MealLog
-import dev.supergooey.caloriesnap.MealLogDatabase
+import dev.supergooey.pical.MealLog
+import dev.supergooey.pical.MealLogDatabase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
