+++ app/src/main/java/dev/supergooey/pical/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package dev.supergooey.caloriesnap.ui.theme
+package dev.supergooey.pical.ui.theme
 
 import androidx.compose.ui.graphics.Color
