+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">CalorieSnap</string>
+    <string name="app_name">Pical</string>
 </resources>
